+++ samples/Lodestar.Sample/Program.cs
@@ -61,6 +61,16 @@ SimHashSample.Run();
 LshBandingSample.Run();
 LshIndexSample.Run();
 
+GpuContextSample.Run();
+GpuSearchResultSample.Run();
+DeviceEmbeddingMatrixSample.Run();
+TiledCosineTopKSample.Run();
+DeviceSparseMatrixSample.Run();
+TiledSparseDenseProductSample.Run();
+DeviceDenseBlockSample.Run();
+DeviceTextBlockSample.Run();
+BitParallelEditDistanceSample.Run();
+
 Bm25IndexSample.Run();
 Bm25OptionsSample.Run();
 SearchHitSample.Run();